Move new SplitCommand and EditFile util functions to internal/util
