+++ cli/util.go
@@ -254,18 +254,6 @@ func splitString(s string) []string {
 	})
 }
 
-// Split the string into a command and its arguments.
-func splitCommand(s string) (string, []string, error) {
-	cmdAndArgs, err := shlex.Split(s)
-	if err != nil {
-		return "", nil, err
-	}
-
-	cmd := cmdAndArgs[0]
-	args := cmdAndArgs[1:]
-	return cmd, args, nil
-}
-
 func splitCLISubjects(subjects []string) []string {
 	new := []string{}
 
@@ -279,32 +267,6 @@ func splitCLISubjects(subjects []string) []string {
 	}
 
 	return new
-}
-
-// Edit the file at filepath f.
-func editFile(f string) error {
-	rawEditor := os.Getenv("EDITOR")
-	if rawEditor == "" {
-		return fmt.Errorf("set EDITOR environment variable to your chosen editor")
-	}
-
-	editor, args, err := splitCommand(rawEditor)
-	if err != nil {
-		return fmt.Errorf("could not parse EDITOR: %v", rawEditor)
-	}
-
-	args = append(args, f)
-	cmd := exec.Command(editor, args...)
-	cmd.Stdin = os.Stdin
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-
-	err = cmd.Run()
-	if err != nil {
-		return fmt.Errorf("could not edit file %v: %s", f, err)
-	}
-
-	return nil
 }
 
 func natsOpts() []nats.Option {
